Refactor ContentView into modular components
